remove types.ts, localize types to individual files
